"fix" test
normalize the received typed parameter, which might be a `typed`
statement/expression list. not ideal, but it's not entirely clear
whether this should be normalized away or not. Even in the previous
implementation a statement list could be received if multiple statements
were written out; at least this way it's symmetric.