transf: fix inline iterator invocation semantics (#745)
## Summary

Specify in the manual how parameter passing for iterators should work,
and adjust `transf` so that inline iterators work according to the new
specification.

This fixes (in the context of inline iterators):
* arguments to `var` and, depending on the argument expression,
  `openArray` parameters being evaluated on *use* inside the iterator's
  body
* literal `ref` construction expression arguments being evaluated on
  *use* (and thus potentially multiple times)
* `sink` parameters not enforcing ownership (the parameter was not
  always mutable, and it also wasn't destroyed at the end of the
  iterator's scope)
* taking the address of an iterator's parameter resulting in illegal
  code being generated by the code generator (this was dependent on
  the shape of the argument expression) 

Closure iterators used with `for`-loops don't adhere to the new
specification yet, and work the same as they did previously.

## Details

Argument handling is split into two parts:
1. deciding the semantics of what the semantics should be (i.e.
   introduce a full copy, inline at parameter usage, etc.). This only
  takes into consideration the parameter and argument type
2. generating the AST for enforcing the semantics

There are four possible modes of how an argument is passed:
- via forwarding: inline the arguments expression at each parameter
  usage
- by copy: create a full copy of the argument. Used for `sink`, single-
  location view, and mutable `openArray`s types. For views, full copy
  only means copying the view, not the viewed location(s)
- by value: depending on the argument, pass either by name or by shallow
  copy. No full copy must be materialized here
- through temporary: assign the argument to a temporary, and the
  temporary to the local introduced for the parameter. This a workaround
  for `transformConv` eliding to-`openArray` conversions

Except for the forwarding mode, each parameter gets allocated a local
corresponding to it, with all usages of the original parameter symbol in
the inlined body being replaced with the symbol of the local.

Previously, locals (read, temporaries) were only introduced when the
argument is a complex expression (with some additional quirks), such as
calls. The temporaries were initialized via fast assignments, but since
those assignments are turned into normal ones by `mirgen`, this provided
no benefit.

Since the inlining logic now always uses either cursors or views for the
locals corresponding to immutable parameters, efficiency of the produced
code slightly increases in some cases.

One place in the compiler code (the `vmbackend.cpairs` iterator) relied
on the previous forwarding semantics for simple argument expressions -- 
the iterator is changed to use a `var` parameter.

Finally, multiple tests for the changed behaviour, as well as the fixed
issues, are added.

### Misc

* rename the `flatten` procedure from `astgen.nim` to `flattenExpr` and
  move it to `trees.nim`
* move the `canUseView` procedure from `astgen.nim` to the new
  `ast_analysis.nim` module