vmgen: fix double evaluation of `swap` arguments (#761)
## Summary

Fix a bug the VM code generator had where the expression used
as the first argument was always evaluated twice -- it is now
only evaluated once.

## Details

The `lowerSwap` procedure that was used for lowering the `swap`
call into assignments duplicated the first argument expression,
resulting in it being evaluated two times.

Usage of `lowerSwap` is replaced with manually emitting the
bytecode for the assignments, which in addition to fixing the
bug and making the generated code slightly more efficient (the
introduced temporary is freed immediately after the swap now),
also has the benefit of removing both the import dependency on
`lowering` and `vmgen` needing access to an `IdGenerator`.

`lowerSwap` has no further usage sites, and is thus removed.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>