internal: drop `nfBlockArg` usage (#760)
## Summary

`nfBlockArg`, an internal flag used to track whether a block argument,
encoded by the `parser` as a `nkStmtList` is no longer used. This is
largely transparent, except for macros receiving a `typed` argument
block will always be wrapped in a statement list, even if there is only
a single statement.

## Details

Internally the compiler tracked block arguments as statement lists with
an internal flag `nfBlockArg` on the node. This prevented various
normalization within semantic analysis from removing this node. This
flag is no longer used and results in only minor changes to parameter
behaviour within `macros`.

Now, any call to a macro where the block style argument syntax is used,
a statement list node is always passed. Whereas before, singleton
statement list nodes were unwrapped, this is no longer the case. This
is considered acceptable as any non-singleton statement would result in
the macro receiving a statement list, and so this behaviour must be
guarded for regardless.

Within the `parser` `isBlockArg` is no longer tracked and in turn
`nfBlockArg` is not applied to `PNode`s during conversion from the
parsed tree to legacy AST. The flag and field in the `ParsedNodeData`
have been removed.

`semStmtList` now exposes a `collapsed` parameter, which `semExprs` uses
to inform statement list unwrapping, this case only when within a
`concept` matching check. `semOperand` has been updated to check if its
input is a statement list, and if it is, it mimics `semExprs` without
statement unwrapping, such as applying `nfSem` to the semanticized
output.

A described earlier, this change means that `typed` arg blocks will not
be unwrapped, and the `tgettypeinst` was adjusted for this.

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>