sequtils: fix double-evaluation bug with `applyIt` (#751)
## Summary

Use the `mitems` iterator for implementing the template, fixing the bug
where the argument expression passed to the template was evaluated
multiple times.

The change also makes the template more general: the only requirement
for the input now is that it supports being iterated over via
`mitems`; it no longer needs to be index-able by an integer.