sigmatch: fix `typeRel` for objects inheriting from phantom types
Summary
=======

Fix the type-relation (`typeRel`) analysis for object types that have a
phantom type somewhere in the inheritance hierarchy. In the context of
type matching, Different instances of a phantom type (e.g. `Type[int]`
and `Type[int]`) were previously treated as being the same.

This also fixes an overloading bug where `Sub = object of Base[T]`
matched equally well against `Base[T]` and `Sub`, resulting in erroneous
ambiguity errors.

Details
=======

When computing the relation between formal = `tyGenericInstance` and
actual = `tyObject`, the closest base type of `actual` that isn't a
`tyObject` object was compared against formal. This skipping was a
previous attempt at fixing the analysis, but it is incorrect, and is
what caused the ambiguity error.

The incorrect type skipping is removed, and replaced with a sub-type
relationship analysis (the new `isSubtypeOfGenericInstance` procedure)
that considers `tyGenericInstance`. This fixes phantom types being
ignored when matching a non-generic object type (with a phantom type
somewhere in the hierarchy) against an instance of a generic type.

Consider the generic type `Type[T] = object`, where `T` is phantom
type information (i.e., it is not used in `Type`'s body). While both
`Type[int]` and `Type[float]` produce distinct `tyObject` types
(different IDs), comparing them with `sameObjectType` would yield
`true`, as the procedure only performs a symbol ID + type structure
comparison!

The `tinheritance_generic_dispatch.nim` test introduced as part of the
previous attempt at a fix is extended, renamed to
`tphantom_type_as_base`, and moved to the `typeRel` category.