backend: lower modules into procedures (#714)
## Summary

- turn module-level imperative code into a procedure (a module's 'init'
  procedure)
- generate the program entry-point in a mostly backend-agnostic way
- change the destruction order of globals, and make it the same across
  backend
- make initialization behaviour of (module-level) globals the same
  across all backend
- make module initialization order the same across all backends (except
  the IC backend)

Almost all decision making regarding initialization order of globals and
modules is moved out of the code generators, and into a central place,
simplifying the code generators and making it much easier to adjust this
behaviour in the future.

Notable changes in behaviour:
- the correct module initialization order is now the one previously used
  with the C backend
- "lifted" globals (those marked with the `.global` pragma) for a module
  are now always initialized before the module-level ones 
- globals are destroyed on a per-module basis; modules are destroyed in
  the reverse order they were initialized (i.e. their module-level code
  run) in
- the `header` pragma is now respected for globals that don't have an
  initial value specified
- `asm` and `emit` statement cannot be used for emitting top-level
  JavaScript code anymore

## Details

### Lowering a module

Turning a module's imperative code into a procedure is integrated into
the collect pass (`setupModule`); once the full semanticized AST for a
module is present, the parts of the declarative AST that the code
generators still need is wrapped in a standalone statement list, while
the imperative AST is pre-processed and wrapped in a procedure.

Pre-processing means:
- scanning the module-level AST for definitions of globals and
  threadvars, and registering them with the `Module` instance
- changing the owner of all symbols defined in the imperative AST to be
  the wrapped-in procedure

The `var` and `let` sections with the globals cannot be turned into
assignments during the pre-processing. This is because:
- `nkForStmt`s weren't lowered yet
- the `injectdestructors` still needs to know about the definition
  points of globals in nested scopes
- the semantics of an "initializing assignment" cannot be encoded with
  `PNode` AST

So instead, transforming the definitions into assignment is done during
the MIR phase (more on that later). Before wrapping up setup of a
`Module` instance, symbols for all other "module-bound" operators are
created. These are:
- module-data initialization procedure
- module de-init (a.k.a. destructor) procedure
- module pre-init procedure

The module destructor is already filled with destructor calls for
globals defined in the outermost scope.

### MIR phase

Transforming the definitions of globals is implemented as two MIR passes
(`rewriteGlobalDefs`). They're run as follows:
- rewrite definitions of globals defined in the *outermost* scope
- `injectdestructors` pass
- rewrite definitions of globals defined in *nested* scopes

With handling of destruction for globals defined in the outermost scope
now being the responsibility of the earlier module processing,
`injectdestructors` only has to defer destructor calls (the
`globalDestructors` mechanism) for lifted globals. The only reason the
`globalDestructors` still needs to exists is that is that `jsgen`
doesn't yet extract lifted globals prior to running the MIR passes. 

After all MIR passes have run, no definitions for globals are left in
the AST that is output for the init procedures -- they were all replaced
with assignments.

### Code generation

The new `backends` module contains code-generator-agnostic routines that
are used by all orchestrators.

The changes to the code generators / orchestrators can be summarized as:
- emitting / handling the definitions for globals is now the
  responsibility of the orchestrators
- deciding on the content of the 'main' procedure is the responsibility
  of the orchestrators (which delegate this to `generateMain`,
  `generateTeardown`, and `generateMainProcedure`)
- the code generators are no longer responsible for handling destruction
  of globals

It's only possible to generate the code for destroying lifted globals
once all alive ones were discovered, but this is only the case after the
main part of code generation is done, as only then has all alive code
been discovered. For this reason, the orchestrators are currently
responsible for calling `finishDeinit` at the correct point in time.

Additional information and highlighted changes.

#### C backend

`cgen` still needs to append code to a module's init procedure from
outside of the init procedure, meaning that the module-init procedures
still have to be special cased.

The code generation logic around globals simplifies a bit, and
`cgenWriteModules` now no longer alters the C code for a module. Because
'main' procedure generation is handled elsewhere, the call to
`initThreadVarsEmulation` is appended to the end of the `system`
module's data-init procedure.

#### JS backend

A special stack-trace entry is inserted for module-level code when using
the JS backend. In order to keep this behaviour, the init procedure is
marked with a symbol flag (the repurposed `sfMainModule`) that signals
to `genProcBody` that the special stack-trace entry is to be used.

#### VM backend

A lot of the code in `vmbackend` was dedicated to module-level code and
entry-point generation, so the module shrinks down quite a bit.

- re-use the "module closed" order information from `ModuleList`
- use the correct module context when generating code
- remove dedicated support for module-level code generation
- replace the manual entry-point code generation with enqueuing the
  procedure produced by `generateMainProcedure`

#### IC backend

In order to make the backend unification process easier, the IC backend
is turned into a frontend for `backend/cbackend.nim`. It mostly acts as
a rodfile -> `ModuleList` adapter now, but still performs the alive
analysis.

A negative side-effect of this is, that when using the IC backend, the
code generator is now also invoked for modules of which neither the
content nor the set of alive symbols changed, significantly increasing
compile times when nothing changed.

### Other changes

- adjust the JavaScript implementation of `std/sysrand`, now that `emit`
  cannot be used to emit top-level JavaScript code anymore
- mark the `globalsSlot` global in `threadlocalstorage` as `.noInit`,
  which is necessary now that all globals without an initial value are
  default-initialized by the module's init procedure

### Changes to tests

- move the test from `tclosure_issues.nim` into a `knownIssue` test; it
  now fails with all backends
- change the `js/tclosure.nim` test to not rely on `asm` producing top-
  level JavaScript code
- don't import modules unavailable for the current platform in
  `tstyle_imports` and `tstrict_funcs_imports` (necessary now that the
  `header` pragma on globals works properly)
- enable some global-related tests for the VM target