jsgen: fix multiple bugs with `addr`
Summary
=======

Fix incorrect JavaScript code being generated for `addr` operations
involving parameters, `let` locations, or imported/exported variables.
In addition, slightly clean up the address-handling in `jsgen`.

Details
=======

A "fat pointer" (array storing a base and index) needs to be constructed
when taking the address of locations not using indirection (such as
`let` locations), but this was missing, and the variable that represents
the location instead returned directly.

* consider locations not boxed in an array (e.g. `let` locations,
  parameters) in `genSymAddr`
* remove the scattered special handling for taking the address of
  parameters
* introduce the `isBoxedPointer` procedure and use it to replace the
  manual checks
* simplify the address assignment logic in `genVarInit` (made possible
  by `genSymAddr` now working properly)
* enable and unrelated test case in `taddr.nim` that already succeeded
  previously