fix(CI): remove self-executing tests from megatest (#768)
## Summary

Marked `texitsignal` as not joinable to exclude it from the `megatest`
category; this should hopefully limit intermitent CI failures.

## Details

`texitsignal` attempts to execute itself with different parameters and
uses that to branch logic of what to check. This doesn't fit with the
rest of the megatest category.

Also, marked these tests as not `joinable` as they either self-execute
or kickoff compiler processes:

- `texception_message_null_byte`
- `tstaticlib`
- `tstackframes`
- `treadlines`
- `tvmops`

Lastly, disabled `tclose` for macOS testing within the same category, as
it's linux only already to save some CI cycles.

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>