testament: small architectural refactoring (#770)
## Summary

Make `TSpec` a pure description of the test specification and don't
reuse it for storing test output. This moves decision-making regarding
whether or not to run a test out of specification parsing, and slightly
improves `testament`s general architecture.

## Details

* remove the `err` property from `TSpec`
* make `inCurrentBatch` a property of `TTest` (instead of `TSpec`)
* introduce the `CompileOutput` type, which captures the output of a
  compiler invocation -- `callNimCompiler` now returns an instance of it
* move early decision-making about what to do with a test (is it
  disabled?, the specification invalid?, etc.) into the new
  `computeEarly` procedure
* adjust `testSpec` and `testSpecWithNimcache` to use `computeEarly`

Due to `compilerOutputTests` re-using `CompileOutput` (previously
`TSpec`) for temporary storage, `CompileOutput` still stores a
`TResultEnum` instead of just a simple boolean.

Multiple types and globals in `spec.nim` can now be moved to
`testament.nim`, but for the sake of a smaller changeset, this is left
to a follow-up commit.