testament: refactor further (#772)
## Summary

Further refactor the `testament` implementation, with the goal of
making it a bit easier to maintain and extend.

There are two notable changes in behaviour:
- `macosx` is no longer supported for the `disabled` key in test
  specifications; `osx` should be used instead
- for `run` actions, the compiler output checks are now performed
  *first*, instead of *after* all other test output checks

## Details

The overall theme is moving decision making about what to do with test
results and interacting with the backend into `testSpecHelper`.

This is mainly achieved by performing output-to-backend step
(`addResult`) in at the end of `testSpecHelper`, with the various output
validation routines now returning a `TestResult` object instead of
outputting the result themselves. As a side-effect, duplicated logic
(incrementing the `passed` counter) is removed and the validation
routines become easier to re-use.

In addition:
- move timeout handling out of `addResult` and into `testSpecHelper`
- make the shell command for executing the compiler (`cmd`) a property
  of `TestRun` (it was previously misplaced in `CompilerOutput`)
- run compiler output checks for `run` actions before the other checks,
  which removes a duplicated code path and makes the behaviour more
  intuitive
- remove the `macosx` value for the `disabled` key  
- `CompilerOutput` now represents the "success" state with a bool
  instead of a `TResultEnum`

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>