testament: support executing "known issue" tests
Summary
=======

Introduce the `--tryFailing` switch for testament, which enables
execution of tests marked as having known issues (and are thus expected
to fail). A "known issue" test successfully finishing in this mode
counts as a test failure.

The plan is to enable this new switch in the CI, in order to
automatically notify about tests that start working after a change. It
is not yet enabled in the CI yet, however.

Details
=======

* add the `--tryFailing` CLI switch, which sets the `runKnownIssue`
  execution flag
* add the `reUnexpectedSuccess` test result code
* render expectedly failing or untried "known issue" tests as
  `KNOWNISSUE` instead of `SKIP` or `NOTINBATCH` in the output