support `method`s with the VM backend (#775)
## Summary

Enable dispatcher generation for the VM backend, which makes `method`
(including multi-methods) available for the standalone VM target. Note
that this doesn't apply to compile-time execution, where using `method`s
is still unsupported.

## Details

* move the `chckNilDisp` compilerproc into the `system` module and
  replace the `== nil` with `isNil` (more on that later)
* only report errors on method usage from `vmgne` when not in the
  `standalone` mode
* rename the `realType` procedure to `routineSignature`
* generate the method dispatcher prior to generating code in
  `vmbackend`
* where possible, enable the `method` tests for the VM target

The VM code generator (`vmgen`) previously checked for whether a
procedure returns something by testing for the presence of a procedure's
`result` AST slot. For methods - where the AST always has a `result`
slot, regardless of whether they have a return value - this test is
no longer enough, and is replaced with inspecting the return type.

Regarding the `== nil` -> `isNil` change: the VM comparison operator for
pointer-like types (which `==` maps to) requires both operands to be of
same dynamic type, but in `chckNilDisp`, this is not necessarily the
case (for example, when the dispatched-through type is a `ref` type).
The proper solution would be for `cgmeth.genDispatcher` to wrap the
argument in a `cast[pointer]`, but the VM doesn't yet support casting
references to pointers.