cgen: revise the interface
Adjust the public interface of the code generator to support the soon-
to-be needs of the orchestrator.

`genProcAux` is split up into `startProc` and `finishProc`, with
`finishProc` returning the generated code instead of emitting it
into a C module section. The `genProc` procedure is kept as a
convenience wrapper around both, and expects .

The init procedure is, for the most part, finally going to be treated
as any other procedure, so all the dedicated code generation logic
(`genInitCode`, `genInitProc`) is removed. The leftover handling of
top-level is also going to be obsolete, and, as a preparation, is
thus removed.

Together, this removes all direct interactions with the MIR and
`transf` from `cgen`.