cgen: prepare for new-style DCE
Don't generate code for a procedure or constant on its first usage.
Instead, only emit a prototype and setup the `TLoc` -- generating the
code for the procedures and constants is going to be the responsibility
of the orchestrator.

Also remove the "is declared" guard for definitions of globals; this
prevents erroneous re-definitions from going unnoticed (they will
cause a C compiler error instead).