cgen: remove `initProc`
A module's 'init' procedure is now treated the same as any other
procedure, with no special assembling by the code generator required.

The last remaining user of the `initProc` field is the `.dynlib`
procedure handling, which is changed to emit manual calls into the
dynlib-init section (which is part of the data-init procedure) instead.
In the future, generating the code for loading `.dynlib` procedures is
going to happen as part of the `process` iterator.