cbackend: fix `.dynlib` foreign dependency discovery
The symbol of the `.dylib` procedure, instead of the foreign dependency,
was registered as the late late-dependency.

In addition, `registerLate` is replaced with adding to the
`hooks` list. The call to `registerLate`, which modifies a variable that
is passed by let-binding to the enclosing for-loops iterator, is
illegal, but the compiler doesn't complain yet.