sempass2: disallow built-ins as procvars in tuples (#783)
## Summary

Disallow using built-in as procvars in the context of literal tuple
constructor expressions, fixing an inconsistency (using built-ins as
procvars is already disallowed everywhere else) that lead to unexpected
behaviour at run-time.

Previously, the code compiled, but, depending on the built-in, the
procvar stored a procedure that did nothing.

## Details

The `nkTupleConstr` processing in `sempass2.track` was missing a call to
`notNilCheck` (which calls `procVarCheck`).

A regression test is added, and a test that made sure the erroneous
expression compiled adjusted.