cgen: split parameter setup from rendering
Summary
=======

Separate `TLoc` setup for parameters from the rendering logic. This
reduces the amount of side-effects the rendering has, and is a
preparation for separating `TLoc` from `TSym`.

Details
=======

* move `TLoc` setup for parameters into `prepareParameters`
* inline the single use of `paramStorage`, and use `mapType` for the
  check
* in `genProcParams`, don't re-compute the information now provided by
  the parameters' `TLoc`
* add calls to `prepareParameters` where needed

This also fixes a regression where a procedure's parameter has no
filled-in `TLoc` if the procedure had not prototype emitted previously.