cgen: fix regression with `.nodecl` consts
Summary
=======

Fix a regression where the `.nodecl` pragma was ignored by the C code
generator in the context of constants.

Details
=======

Apart from the fix, the single-use `genConstSetup` is inlined into
`useConst`. Since the loc for constants is setup in `genConstDefinition`
(which is always called before `useConst`), the `fillLoc` call is not
added to `useConst` (this is also a step towards `fillLoc` only being
used in procedures responsible for definitions).

In addition, a test for the current behaviour of imported `.nodecl`
constants is added.