cgen: stricter symbol handling for `emit`
Summary
=======

Slightly refactor `ccgstmts.genAsmOrEmitStmt` and make its symbol
handling logic more strict. Apart from fields and types, all symbols
are now required to have their loc name already generated.

This is only an internal change. It prevents regressions, makes the
logic more clear, and is also a preparation for removing `TLoc` from
`TSym`.

Details
=======

An exhaustive case statement is used instead of the 'if', and the
redundant `$` calls are removed.

Due to the lack of context, field names cannot always be properly
mangled when the field is missing a loc name, but instead of the generic
name mangling (original name + attached-to module name + item id) the
record field mangling scheme is now used (which is incorrect for tuple
fields, but `mangleName` was too).