internal: separate `TLoc` from `TSym` and `TType` (#790)
## Summary

Remove the `loc` fields from both `TSym` and `TType` and leave it to
the code generators how and if they use `TLoc` -- `TSym` only stores the
external name plus interface flags now. The benefits:
- another point of coupling (through data-types) between semantic
  analysis and code generation is removed
- the size of an `TType` instance shrinks from 120 to 96 bytes (with a
  64-bit target)
- the size of an `TSym` instance shrinks from 168 to 160 bytes

Because of its `lode` field, `TLoc` not being defined in the `ast_types`
module is also an important prerequisite for introducing a code IR for
the code generators.

## Details

The key changes:
- move `TLoc` and the associated types only relevant to the C code
  generator to `cgendata`
- remove the `loc` field from `TSym` and `TType` -- `TSym` stores the
  external name and interface flags via the new `extname` and `locFlags`
  fields
- as an interim solution, add the `locId` field to `TSym`
- fix the name slot in the `ast` of dispatcher storing the wrong symbol
- NDI file generation is now the responsibility of the orchestrator

### `cgen`

For associating a `PSym` with a `TLoc`, multiple `Store`s shared across
all C modules are used. The `TSym.locId` is an index into the
`SymbolMap` (which is a `Store` underneath) for the respective symbol
kind (globals, constants, procedures, fields). While a single
`SymbolMap` could be used (for the non-procedure entities), the plan is
to use dedicated types for the entities in the future, and using
separate stores is a preparation for that.

In order to not re-mangle parameters every time when emitting a
prototype, the locs for parameters are stored with their procedure.

Since locals are known to not outlive their surrounding procedure, the
locs for them are stored in a dedicated `SymbolMap` with the procedure
context -- once a procedure context goes away, the locals' `TLoc`s can
be, and are, freed already.

In general, the C code generator is now more strict with who is
responsible for filling-in loc information: for constants and globals,
it must only happen once during definition. Procedures and fields still
need to support ad-hoc loc setup, and fields need special-casing because
of `.inline` procedures and how `finally` clauses work (i.e., they're
duplicated at a too late time).

A table-based approach where the symbol IDs are mapped to `TLoc`s would
have also worked, but measurements showed that it was ~1.4% slower.

#### NDI files

Instead of directly updating the NDI file with a new entry when
generating a mangled name, as was done previously, adding the entries is
now the responsibility of the orchestrator: entries for whole-program
entities are added after code generation has finished -- entries for
locals once code generation for a procedure has finished.

For NDI file generation to work, the symbol. This was already the case
for all entities except constants, so constants now also store a `nkSym`
node in their `lode` field.

### `jsgen`

Due to it being less complex than the C code generator, a table-based
approach is okay for the JavaScript code generator. Only the JavaScript
name is relevant to the code generator, so it maps the symbol IDs
directly to strings and doesn't use `TLoc`. Separate tables are used
for global entities (globals, constants, procedures, and fields),
locals, and parameters.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>