internal: split up `TLocFlag`
Summary
=======

Separate the enum into two enums: one for the flags describing a
symbol's external interface and one for the actual location flags. This
fixes an implementation detail of the C code generator (location flags)
leaking into `TSym`.

Details
=======

The enum for the external interface flags is named `ExternalFlag` and
the location flag enum `LocFlag`. Since the latter is only relevant to
the C code generator, it is moved to `cgendata`.

The field storing the external interface flags in `TSym` and `PackedSym`
is renamed to `extFlags`. Due to the `ef` prefix being already taken by
the `TExprFlag` enum, the `ExternalFlag` enum uses the `exf` prefix.