move the MIR generation DSL to `mirconstr`
It's self-contained and doesn't depend on the rest of `mirgen`. The
`notOp`, `modify`, `outOp`, and `tupleAccess` operators are not moved --
they're not atoms like the others.

For convenience and to keep the set of changes smaller, `mirgen` uses
wrapper templates around some of the DSL routines.

Apart from some very small documentation fixes, the moved routines
are not modified.