cgen: replace `TLoc` for procs and fields
Summary
=======

Use dedicated data types (instead of `TLoc`) for storing C-related
information about procedures and fields in `cgen`. This slightly reduces
the memory usage, and it also makes the types easier to evolve.

Details
=======

For fields, only the mangled (or unmangled) name is relevant, all other
information stored by `TLoc` is not. Therefore, storing only a string is
enough. In addition, the `fieldLoc` procedure is renamed to `fieldName`.

For procedures, the only information that needs to be stored is the name
and the parameters' loc. As a workaround for NDI file writing needing a
`PSym` to fetch the user-provided name from, the new `ProcLoc` type also
stores the `PSym`. Once each `ProcLoc` entry is associated with a
`DiscoveryData.procedures` entry through its index, this workaround can
be removed.